refactor(ts/strands): split strands.ts into per-construct modules
846 lines with three classes and a factory, about to gain a memory store.
Split by construct; index.ts re-exports the identical public surface and
becomes the tsup entry target, so the published specifier and the emitted
dist path are unchanged. No behaviour change.
