+++ typescript/src/integrations/strands/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Strands Agents integration. Split across modules by construct; this file
+ * is the only public entry point and the tsup entry target, so the
+ * `@neo4j-labs/agent-memory/integrations/strands` specifier is unchanged.
+ */
+export type { StrandsIntegrationOptions } from "./internal.js";
+export {
+  isSyntheticStrandsMessage,
+  SYNTHETIC_MESSAGE_PREFIXES,
+} from "./synthetic.js";
+export { Neo4jSessionStorage } from "./session-storage.js";
+export {
+  Neo4jConversationManager,
+  type Neo4jConversationManagerOptions,
+} from "./conversation-manager.js";
+export { registerReasoningHooks, type ReasoningHooksOptions } from "./reasoning-hooks.js";
+export { connectMemoryToAgent, type ConnectMemoryToAgentResult } from "./connect.js";
