+++ typescript/test/integration/strands-hooks.test.ts
@@ -15,7 +15,7 @@ import {
   type LocalAgent,
 } from "@strands-agents/sdk";
 import { MemoryClient } from "../../src/client.js";
-import { registerReasoningHooks } from "../../src/integrations/strands.js";
+import { registerReasoningHooks } from "../../src/integrations/strands/index.js";
 
 const ENDPOINT = "https://memory.test/v1";
 const API_KEY = "nams_test_key";
