+++ typescript/tsup.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
     "mcp/index": "src/mcp/index.ts",
     "integrations/langchain": "src/integrations/langchain.ts",
     "integrations/mastra": "src/integrations/mastra.ts",
-    "integrations/strands": "src/integrations/strands.ts",
+    "integrations/strands": "src/integrations/strands/index.ts",
     testing: "src/testing.ts",
   },
   format: ["esm"],
